test(codex): isolate response boundary fixture
